Reorg code into smaller files, save/ load config
Resolves #32
Resolves #25

Functionality is split into clear and maintainable files, parameter logic is looking very strong, and is more based on EEPROM values than ever before, menuBar now has a Tools section with a config dump, and it is roughly possible to save and load config files (though this functionality will be improved). Finally, functions and variables have been renamed here and there for consistency
