+++ Software/ClientApp/src/main/python/gui_elements/version.py
@@ -1,4 +1,4 @@
-__version__ = 'v1.3.0'
-__date__ = "20221006"
+__version__ = "v1.9.9"
+__date__ = "20221010"
 __appname__ = "PoT_Config_Tool"
-__appid__ = 'hld.PoTConfig' + __version__
+__appid__ = "hld.PoTConfig" + __version__
